storage size set to biginteger
Bug-Url: https://github.com/K0zka/kerub/issues/88
Bug-Url: https://github.com/K0zka/kerub/issues/23

+++ src/main/kotlin/com/github/K0zka/kerub/model/VirtualStorageDevice.kt
@@ -3,6 +3,7 @@ package com.github.K0zka.kerub.model
 import com.fasterxml.jackson.annotation.JsonTypeName
 import org.hibernate.search.annotations.DocumentId
 import org.hibernate.search.annotations.Field
+import java.math.BigInteger
 import java.util.UUID
 
 JsonTypeName("vstorage")
@@ -10,7 +11,7 @@ data class VirtualStorageDevice(
 		DocumentId
 		override val id: UUID = UUID.randomUUID(),
 		Field
-		val size: Long,
+		val size: BigInteger,
 		Field
 		val readOnly: Boolean = false,
 		Field
