+++ src/main/kotlin/com/github/K0zka/kerub/model/dynamic/VirtualMachineDynamic.kt
@@ -1,6 +1,7 @@
 package com.github.K0zka.kerub.model.dynamic
 
 import com.github.K0zka.kerub.model.VirtualMachineStatus
+import java.math.BigInteger
 import java.util.UUID
 
 public data class VirtualMachineDynamic(
@@ -9,6 +10,6 @@ public data class VirtualMachineDynamic(
 		override val lastUpdated: Long = System.currentTimeMillis(),
         val hostId : UUID,
         val status : VirtualMachineStatus = VirtualMachineStatus.Down,
-        val memoryUsed : Long
+        val memoryUsed : BigInteger
                                        ) : DynamicEntity {
 }
