+++ src/main/kotlin/com/github/K0zka/kerub/planner/costs/NetworkCost.kt
@@ -2,4 +2,4 @@ package com.github.K0zka.kerub.planner.costs
 
 import com.github.K0zka.kerub.model.Host
 
-data class NetworkCost(bytes : Int, hosts : List<Host>) : Cost
+data class NetworkCost(bytes : Long, hosts : List<Host>) : Cost
